Complete rework DTOs are now one per entity. Services and controllers deleted to start from scratch. Question controller GET, POST, PUT, DELETE mappings working.

+++ src/main/java/org/wowyomad/questionaire/repository/OptionRepository.java
@@ -1,10 +1,10 @@
 package org.wowyomad.questionaire.repository;
 
 
-import org.wowyomad.questionaire.model.Option;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.wowyomad.questionaire.model.Option;
 
-public interface OptionRepository extends JpaRepository<Option, Long> {
+public interface OptionRepository extends JpaRepository<Option, Integer> {
 
 }
