+++ src/main/java/org/wowyomad/questionaire/repository/QuestionRepository.java
@@ -3,6 +3,6 @@ package org.wowyomad.questionaire.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.wowyomad.questionaire.model.Question;
 
-public interface QuestionRepository extends JpaRepository<Question, Long> {
+public interface QuestionRepository extends JpaRepository<Question, Integer> {
 
 }
