+++ src/main/java/org/wowyomad/questionaire/repository/SubmissionRepository.java
@@ -1,8 +1,8 @@
 package org.wowyomad.questionaire.repository;
 
-import org.wowyomad.questionaire.model.Submission;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.wowyomad.questionaire.model.Submission;
 
-public interface SubmissionRepository extends JpaRepository<Submission, Long> {
+public interface SubmissionRepository extends JpaRepository<Submission, Integer> {
 }
